requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
anthropic
genruntime
httpx